+++ upcloud/config.go
@@ -46,7 +46,7 @@ func (c *Config) checkLogin(svc *service.Service) (*upcloud.Account, error) {
 		if ok {
 			return nil, fmt.Errorf("[ERROR] Failed to get account, error was %s: '%s'", svcErr.ErrorCode, svcErr.ErrorMessage)
 		}
-		return nil, fmt.Errorf("[ERROR] Failed to get account due to unspecified error")
+		return nil, err
 	}
 
 	return res, nil