+++ CHANGELOG.md
@@ -5,9 +5,14 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- storage: upcloud_storage data source to retrieve specific storage details
+
 ### Fixed
 
 - docs: set provider username and password as required arguments
+- provider: return underlying error from initial login check instead of custom error
 
 
 ## [2.1.5] - 2022-01-27